refactor(memory): split consolidation into candidate, signals, deep-ranking, markers, plan, and diary modules
Behavior-preserving along the file's existing seams; consolidation.ts names
the parts together for the callers that read consolidation whole, and the
barrel is unchanged. Along the way: the wire reader uses the wire package's
own field readers, the marker back-walk is one removeEntryWithMarkers, the
importance rank drops its unused parameter, recall signals import the notes
directory rather than defaulting its name, and DAY_MS is one constant in
runtime-contracts.
