+++ packages/runtime-contracts/src/index.ts
@@ -189,4 +189,4 @@ export {
   TRANSCRIPT_EVENT_KIND,
   type TranscriptEvent,
 } from "./storage.js";
-export type { ScheduledTimer } from "./timers.js";
+export { DAY_MS, type ScheduledTimer } from "./timers.js";
