+++ packages/runtime-contracts/src/timers.ts
@@ -6,3 +6,6 @@
  * package is cancellable in another.
  */
 export type ScheduledTimer = number | object;
+
+/** One day in milliseconds, for every age, half-life, and lookback measured in days. */
+export const DAY_MS = 24 * 60 * 60 * 1000;
